added simple poc assembly import/export scripts
git-svn-id: http://vcs.taotesting.com/svn/oat/projects/taoDelivery/trunk@20662 62dfeb86-a93f-4914-9d13-15c0a160985d

+++ models/classes/import/class.Assembler.php
@@ -65,7 +65,9 @@ class taoDelivery_models_classes_import_Assembler
             ));
             $report = common_report_Report::createSuccess(__('Delivery "%s" successfully imported',$label), $delivery);
         } catch (Exception $e) {
-            $delivery->delete();
+            if (isset($delivery) && $delivery instanceof core_kernel_classes_Resource) {
+                $delivery->delete();
+            }
             $report = common_report_Report::createFailure(__('Unkown error during impoort'));
         }
         return $report;
